% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myData.R
\docType{data}
\name{myData}
\alias{myData}
\title{Investor data}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{Y}{the percentage of total wealth either gained or lost in 2018 from these stocks}
\item{X1}{the percentage of total wealth either gained from selling or lost from buying stock in Disney}
\item{X2}{the percentage of total wealth either gained from selling or lost from buying stock in Apple}
\item{X3}{the percentage of total wealth either gained from selling or lost from buying stock in Tesla}
\item{X4}{the percentage of total wealth either gained from selling or lost from buying stock in Microsoft}
\item{X5}{the percentage of total wealth either gained from selling or lost from buying stock in Airbnb}
...
}
}
\source{
\url{https://canvas.vt.edu/courses/108765/files?preview=13456230}
}
\usage{
myData
}
\description{
This data shows how 100 investors' wealth either increased or decreased
as a percentage of their total wealth (column "Y") based on what percentage
of their total wealth they either sold stock (positive number) or bought
stock (negative number) in any one of 5 major stocks in 2018.
}
\keyword{datasets}
